backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
google-generativeai==0.3.1
elevenlabs==0.2.27
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pydantic==2.5.0
